SpringValidator para correo y contraseña
JPA para Productos (Vacios aun, recordar llenar la base de datos)

Mocks aun deshabilitados

+++ backend/src/main/java/com/cursoIntegrador/lePettiteCoffe/Repository/AccountRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.cursoIntegrador.lePettiteCoffe.Model.Entity.Cuenta;
 
 @Repository
-public interface AccountRepository extends JpaRepository<Cuenta, Long> {
+public interface AccountRepository extends JpaRepository<Cuenta, Integer> {
 
     Cuenta findByEmail(String email);
